Refactor whole polyglote crate
